# api/openapi.yaml
openapi: 3.1.2
info:
  title: AXON API
  version: 2026-07-14
  description: Build production AI agents without building the infrastructure. Bring your own model.
servers:
  - url: http://127.0.0.1:4318
paths:
  /health:
    get:
      operationId: health
      responses:
        '200': { description: Service is healthy }
  /v1/agents:
    post:
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentRequest' }
      responses:
        '201': { description: Agent created, content: { application/json: { schema: { $ref: '#/components/schemas/Agent' } } } }
        '400': { $ref: '#/components/responses/Error' }
  /v1/agents/{agent_id}/run:
    post:
      operationId: runAgent
      parameters: [{ $ref: '#/components/parameters/AgentID' }]
      responses:
        '202': { description: Run queued, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /v1/agents/{agent_id}/pause:
    post:
      operationId: pauseAgent
      parameters: [{ $ref: '#/components/parameters/AgentID' }]
      responses: { '200': { description: Agent paused } }
  /v1/agents/{agent_id}/resume:
    post:
      operationId: resumeAgent
      parameters: [{ $ref: '#/components/parameters/AgentID' }]
      responses: { '200': { description: Agent resumed } }
  /v1/agents/{agent_id}/logs:
    get:
      operationId: listAgentLogs
      parameters: [{ $ref: '#/components/parameters/AgentID' }]
      responses: { '200': { description: Agent logs } }
components:
  parameters:
    AgentID: { name: agent_id, in: path, required: true, schema: { type: string } }
  responses:
    Error: { description: API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    CreateAgentRequest:
      type: object
      required: [workspace, name, goal]
      properties:
        workspace: { type: string }
        name: { type: string }
        goal: { type: string }
    Agent:
      type: object
      required: [id, workspace, name, goal, status, created_at]
      properties:
        id: { type: string }
        workspace: { type: string }
        name: { type: string }
        goal: { type: string }
        status: { type: string, enum: [active, paused] }
        created_at: { type: string, format: date-time }
    Run:
      type: object
      required: [id, agent_id, status, created_at]
      properties:
        id: { type: string }
        agent_id: { type: string }
        status: { type: string, enum: [queued] }
        created_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [type, code, message, request_id]
          properties:
            type: { type: string }
            code: { type: string }
            message: { type: string }
            request_id: { type: string }